requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
bcrypt